Lib

Documentation

lt :: (Ord a, Show a) => a -> a -> Property

mapMonotonicallyInc can reduce the needed number of required applications of the given function. For example, consider the following case:

 mapMonotonicallyInc :: Set a -> Set ()
 mapMonotonicallyInc (const ())

while S.map (const ()) takes O(n log n) time, the above takes the ideal: O(1) constant time.

The following case also exhibits this behavior:

 mapMonotonicallyInc ceiling $
   fromDistinctAscList
     [0.64, 0.75, 0.95]

Property that one is less than another, with counterexample

leq :: (Ord a, Show a) => a -> a -> Property

Property that one is less than or equal to another, with counterexample

(/==) :: (Eq a, Show a) => a -> a -> Property infix 4

Property that the arguments are non-equal, with counterexample

sumNotMonoInc :: ()

The sum of a set is not monotonically increasing:

 fromList [1,4]
 fromList [2]
 5 >  2

fixMaybeOn :: (a -> Maybe a) -> a -> a

Apply the function repeatedly until it returns Nothing. Return the last Just result or the input if there is none.

resize :: a -> Set a -> Set a -> Set a

Recalculate the size given a root and its two immediate sub-Sets

resizeF :: (Int -> a -> Set a -> Set a -> b) -> a -> Set a -> Set a -> b

Recalculate the size a la resize, also given a constructor to produce the result

joinLeft :: Eq a => Set a -> Maybe (Set a)

If the root of the tree is equal to the root of the left subtree, then so are the elements of the left subtree's right subtree. We join them with their new sizes.

joinRight :: Eq a => Set a -> Maybe (Set a)

See joinLeft

firstJoins :: Eq a => Set a -> Set a

joinRight as many times as possible, then do the same with joinLeft

joinLrs :: Eq a => Set a -> Set a

Reimplementation of firstJoins

joinRights :: Eq a => Set a -> Set a

Join equal values along right sub-Sets

joinRightsLoop :: Eq a => a -> Set a -> Set a -> Set a

joinRights on a known non-Tip Set

joins :: Eq a => Set a -> Set a

Join equal elements

joinsF :: Eq a => a -> Set a -> Set a -> Set a

joins on a known non-Tip Set

joins'' :: Eq a => Set a -> Set a

mapLevels firstJoins

joins' :: Eq a => Set a -> Set a

mapLevels joinLrs

mapLevels :: (Set a -> Set a) -> Set a -> Set a

mapMonotonicallyIncreasing :: Eq b => (a -> b) -> Set a -> Set b

Using joins

mapMonotonicallyIncreasing' :: Eq b => (a -> b) -> Set a -> Set b

Using joins'

mapMonotonicallyIncreasing'' :: Eq b => (a -> b) -> Set a -> Set b

Using joins''

impl :: Bool -> Bool -> Bool

Logical implication

uncons :: Ord a => Set a -> Maybe (a, Set a)

Get Just the first element of a Set and the rest of it or Nothing

mapMonotonicHolds :: (Ord a, Ord b, Show a, Show b) => ((a -> b) -> Set a -> Set b) -> (a -> b) -> Set a -> Property

mapMonotonicallyIncreasing with duplicates removed is equivalent to map for monotonically increasing functions

mapMonotonicHoldsSubset :: (Ord a, Show a) => ((a -> a) -> Set a -> Set a) -> Set a -> Set a -> Property

mapMonotonicHolds where the first element of a Set replaces all elements not in the first Set and the rest are untouched.

newtype Stream a

Infinite streams

Constructors

Stream

Fields

runStream :: Cofree Identity a

subsetBy :: Ord a => Stream Bool -> Set a -> Set a

Keep in Set if True

withSubset :: Ord a => (Set a -> Set a -> b) -> Stream Bool -> Set a -> b

Use subsetBy to provide the first argument

type SubsetProp a = Stream Bool -> Set a -> Property

Subset-input properties

prop_mapMonotonicHoldsSubsetWord4 :: SubsetProp Word4

mapMonotonicHoldsSubset mapMonotonicallyIncreasing for Word4

prop_mapMonotonicHoldsSubsetWord8 :: SubsetProp Word8

mapMonotonicHoldsSubset mapMonotonicallyIncreasing for Word8

prop_mapMonotonicHoldsSubsetWord4' :: SubsetProp Word4

mapMonotonicHoldsSubset mapMonotonicallyIncreasing' for Word4

prop_mapMonotonicHoldsSubsetWord8' :: SubsetProp Word8

mapMonotonicHoldsSubset mapMonotonicallyIncreasing' for Word8

prop_mapMonotonicHoldsSubsetWord4'' :: SubsetProp Word4

mapMonotonicHoldsSubset mapMonotonicallyIncreasing'' for Word4

prop_mapMonotonicHoldsSubsetWord8'' :: SubsetProp Word8

mapMonotonicHoldsSubset mapMonotonicallyIncreasing'' for Word8

countSubsetSums :: Int -> Int

countSubsetSums2 :: Int -> Int

countSubsetSums3 :: Int -> Int

setupEnv :: IO (Set Word8, Set Word16, Set Word32)

Set up bench environment with:

 let small = [1..100]
     large = [1..2^10]
     halfOverlapping = S.mapMonotonic (x -> if even x then x + 1 else x) [1..2^20]
 return (small, large, halfOverlapping)

runBenches :: IO ()

Run benchmarks

runTests :: IO Bool

Run all prop_ tests